+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 * Added support for fur shell rendering using geometry nodes.
 
+### Fixed
+* Fixed an issue where camdo import would not import all bones in some cases.
+
 ## 0.20.2 - 2025-10-10
 ### Fixed
 * Fixed a potential crash when importing normals for model and map geometry in some cases.