+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## unreleased
+## 0.21.0 - 2025-11-19
 ### Added
 * Added import and export support for fur shells for wimdo models using geometry nodes.